requirements.txt
annotated-types==0.6.0
ansicolors==1.1.8
pydantic==2.6.2
pydantic_core==2.16.3
typing_extensions==4.9.0
